Initial work on PAL
