+++ src/ComputeSharp.Dynamic/ComputeSharp.Dynamic.csproj
@@ -4,7 +4,7 @@
     <OutputPath>bin\Release</OutputPath>
     <DocumentationFile>ComputeSharp.Dynamic.xml</DocumentationFile>
     <Platforms>AnyCPU;x64;ARM64</Platforms>
-    <SupportedOSVersion>windows6.2</SupportedOSVersion>
+    <!-- <SupportedOSVersion>windows6.2</SupportedOSVersion> -->
   </PropertyGroup>
 
   <ItemGroup>
