+++ src/ComputeSharp.WinUI/ComputeSharp.WinUI.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
-    <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
+    <TargetFramework>net6.0</TargetFramework>
+    <!-- <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion> -->
     <RootNamespace>ComputeSharp.WinUI</RootNamespace>
     <Platforms>x64;ARM64</Platforms>
     <RuntimeIdentifiers>win10-x64;win10-arm64</RuntimeIdentifiers>
