+++ src/ComputeSharp/ComputeSharp.csproj
@@ -3,7 +3,7 @@
     <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
     <OutputPath>bin\Release</OutputPath>
     <DocumentationFile>ComputeSharp.xml</DocumentationFile>
-    <SupportedOSVersion>windows6.2</SupportedOSVersion>
+    <!-- <SupportedOSVersion>windows6.2</SupportedOSVersion> -->
   </PropertyGroup>
 
   <ItemGroup>
